# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AlertTimelineItem:
    properties:
      date:
        description: 日期
        type: string
      firing:
        description: 触发数量
        type: integer
      resolved:
        description: 恢复数量
        type: integer
    type: object
...
  dto.CreateRequirementPoolRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      owner_id:
        type: integer
      project_id:
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/model.RequirementPoolType'
        enum:
        - global
        - project
    required:
    - name
    - owner_id
    - type
    type: object
...
  dto.DailyWorklogStat:
    properties:
      date:
        type: string
      entry_count:
        type: integer
      total_time_sec:
        type: integer
    type: object
  dto.DashboardStatsResponse:
    properties:
      alert_stats:
        description: 告警统计
        properties:
          today_created:
            description: 今日告警
            type: integer
          total_acked:
            description: 已确认告警
            type: integer
          total_firing:
            description: 触发中告警
            type: integer
          total_resolved:
            description: 已恢复告警
            type: integer
          week_created:
            description: 本周告警
            type: integer
        type: object
      issue_stats:
        description: 工单统计
        properties:
          today_created:
            description: 今日创建
            type: integer
          total_closed:
            description: 已关闭工单
            type: integer
          total_in_progress:
            description: 进行中工单
            type: integer
          total_open:
            description: 待处理工单
            type: integer
          total_resolved:
            description: 已解决工单
            type: integer
          week_created:
            description: 本周创建
            type: integer
          week_resolved:
            description: 本周解决
            type: integer
        type: object
      project_stats:
        description: 项目统计
        properties:
          total_members:
            description: 总成员数
            type: integer
          total_projects:
            description: 总项目数
            type: integer
        type: object
    type: object
...
  dto.DistributionItem:
    properties:
      name:
        description: 名称
        type: string
      ratio:
        description: 占比
        type: number
      value:
        description: 数量
        type: integer
    type: object
...
  dto.IssueStatsResponse:
    properties:
      assignee_distribution:
        description: 按指派人分布
        items:
          $ref: '#/definitions/dto.DistributionItem'
        type: array
      epic_distribution:
        description: 按 Epic 分布
        items:
          $ref: '#/definitions/dto.DistributionItem'
        type: array
      priority_distribution:
        description: 按优先级分布
        items:
          $ref: '#/definitions/dto.DistributionItem'
        type: array
      status_distribution:
        description: 按状态分布
        items:
          $ref: '#/definitions/dto.DistributionItem'
        type: array
      summary:
        properties:
          avg_resolve_time:
            description: 平均解决时间（小时）
            type: number
          closed:
            description: 已关闭
            type: integer
          in_progress:
            description: 进行中
            type: integer
          open:
            description: 待处理
            type: integer
          resolved:
            description: 已解决
            type: integer
          total:
            description: 总数
            type: integer
        type: object
      timeline:
        description: 按时间分组的统计
        items:
          $ref: '#/definitions/dto.TimelineItem'
        type: array
      type_distribution:
        description: 按类型分布
        items:
          $ref: '#/definitions/dto.DistributionItem'
        type: array
    type: object
...
  dto.RequirementPoolResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      owner_id:
        type: integer
      owner_name:
        type: string
      project_id:
        type: integer
      project_name:
        type: string
      requirement_count:
        description: 需求数量
        type: integer
      status:
        $ref: '#/definitions/model.RequirementPoolStatus'
      type:
        $ref: '#/definitions/model.RequirementPoolType'
      updated_at:
        type: string
    type: object
...
  dto.SLAPriorityStats:
    properties:
      mtta:
        description: 平均确认时间
        type: number
      mttr:
        description: 平均解决时间
        type: number
      priority:
        description: 优先级
        type: string
      resolved:
        description: 已解决
        type: integer
      sla_met:
        description: 达标数
        type: integer
      sla_rate:
        description: 达标率
        type: number
      sla_target:
        description: SLA 目标（分钟）
        type: integer
      total:
        description: 总数
        type: integer
    type: object
  dto.SLAProjectStats:
    properties:
      mttr:
        description: 平均解决时间
        type: number
      project_key:
        description: 项目 Key
        type: string
      project_name:
        description: 项目名称
        type: string
      resolved:
        description: 已解决
        type: integer
      sla_rate:
        description: 达标率
        type: number
      total:
        description: 总数
        type: integer
    type: object
  dto.SLAReportResponse:
    properties:
      by_priority:
        description: 按优先级的 SLA 统计
        items:
          $ref: '#/definitions/dto.SLAPriorityStats'
        type: array
      by_project:
        description: 按项目的 SLA 统计
        items:
          $ref: '#/definitions/dto.SLAProjectStats'
        type: array
      summary:
        properties:
          mtta:
            description: 平均确认时间（分钟）
            type: number
          mttr:
            description: 平均解决时间（分钟）
            type: number
          resolved_issues:
            description: 已解决工单数
            type: integer
          sla_met:
            description: SLA 达标数
            type: integer
          sla_rate:
            description: SLA 达标率（%）
            type: number
          sla_violated:
            description: SLA 违规数
            type: integer
          total_issues:
            description: 总工单数
            type: integer
        type: object
      violations:
        description: SLA 违规工单列表
        items:
          $ref: '#/definitions/dto.SLAViolation'
        type: array
    type: object
  dto.SLAViolation:
    properties:
      actual_time:
        description: 实际耗时（分钟）
        type: number
      assignee_name:
        description: 经办人
        type: string
      issue_key:
        description: 工单号
        type: string
      overdue_by:
        description: 超时时长（分钟）
        type: number
      priority:
        description: 优先级
        type: string
      sla_target:
        description: SLA 目标（分钟）
        type: integer
      title:
        description: 标题
        type: string
    type: object
...
  dto.TimelineItem:
    properties:
      closed:
        description: 关闭数量
        type: integer
      created:
        description: 创建数量
        type: integer
      date:
        description: 日期
        type: string
      in_progress:
        description: 进行中数量
        type: integer
      resolved:
        description: 解决数量
        type: integer
    type: object
...
  dto.TopAlertItem:
    properties:
      alert_name:
        description: 告警名称
        type: string
      count:
        description: 数量
        type: integer
      severity:
        description: 严重程度
        type: string
    type: object
...
  dto.UpdateRequirementPoolRequest:
    properties:
      description:
        maxLength: 5000
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      owner_id:
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/model.RequirementPoolStatus'
        enum:
        - active
        - archived
    type: object
...
  dto.UserPerformanceResponse:
    properties:
      assigned:
        description: 指派数
        type: integer
      avg_resolve_time:
        description: 平均解决时间
        type: number
      display_name:
        type: string
      resolved:
        description: 解决数
        type: integer
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  dto.UserWorklogStat:
    properties:
      display_name:
        type: string
      entry_count:
        type: integer
      total_time_sec:
        type: integer
      user_id:
        type: integer
    type: object
...
  dto.WorklogDetail:
    properties:
      issue_key:
        type: string
      time_sec:
        type: integer
      title:
        type: string
    type: object
  dto.WorklogGridRow:
    properties:
      daily:
        additionalProperties:
          format: int64
          type: integer
        description: date -> 秒
        type: object
      daily_details:
        additionalProperties:
          items:
            $ref: '#/definitions/dto.WorklogDetail'
          type: array
        description: date -> 工单明细
        type: object
      display_name:
        type: string
      total_sec:
        description: 该用户总工时（秒）
        type: integer
      user_id:
        type: integer
    type: object
...
  dto.WorklogStatsResponse:
    properties:
      daily_stats:
        items:
          $ref: '#/definitions/dto.DailyWorklogStat'
        type: array
      grid:
        description: 用户×日期工时明细
        items:
          $ref: '#/definitions/dto.WorklogGridRow'
        type: array
      grid_dates:
        description: 有序日期列表
        items:
          type: string
        type: array
      grid_totals:
        additionalProperties:
          format: int64
          type: integer
        description: 每日合计（秒）
        type: object
      summary:
        $ref: '#/definitions/dto.WorklogSummary'
      type_stats:
        items:
          $ref: '#/definitions/dto.WorklogTypeStat'
        type: array
      user_stats:
        items:
          $ref: '#/definitions/dto.UserWorklogStat'
        type: array
    type: object
  dto.WorklogSummary:
    properties:
      active_users:
        type: integer
      avg_daily_time_sec:
        type: number
      total_entries:
        type: integer
      total_time_sec:
        type: integer
    type: object
  dto.WorklogTypeStat:
    properties:
      entry_count:
        type: integer
      total_time_sec:
        type: integer
      work_type:
        type: string
    type: object
...
  internal_reporting_dto.AlertStatsResponse:
    properties:
      severity_distribution:
        description: 按严重程度分布
        items:
          $ref: '#/definitions/dto.DistributionItem'
        type: array
      summary:
        properties:
          acked:
            description: 已确认
            type: integer
          avg_ack_time:
            description: 平均确认时间（分钟）
            type: number
          firing:
            description: 触发中
            type: integer
          resolved:
            description: 已恢复
            type: integer
          total:
            description: 总数
            type: integer
        type: object
      timeline:
        description: 按时间分组的统计
        items:
          $ref: '#/definitions/dto.AlertTimelineItem'
        type: array
      top_alerts:
        description: 按告警名称 Top10
        items:
          $ref: '#/definitions/dto.TopAlertItem'
        type: array
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.PageData:
    properties:
      has_more:
        description: true 表示实际总数超过 Total，前端显示 "10,000+"
        type: boolean
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:10010
info:
  contact:
    name: TicketDesk
    url: https://github.com/kerbos/ticketdesk
  description: |-
    项目化工单与告警联动系统 API. 一切问题都是工单, 一切告警都必须被跟进.
    **统一响应格式**: `{code, message, data}`. 错误响应 `{code, message, details?}` (code 为字符串如 "BAD_REQUEST" / "UNAUTHORIZED").
    **速率限制**: 全站默认 300 req/min/IP, /auth/* 端点 20 req/min/IP, /webhook/* 100 req/min/IP. 超出返回 **429 Too Many Requests**, 此状态码在所有端点皆可能出现, 单独 endpoint 文档不重复声明.
  license:
    name: MIT
  termsOfService: https://github.com/kerbos/ticketdesk
  title: TicketDesk API
  version: "1.0"
paths:
...
  /api/v1/reports/alerts:
    get:
      description: 获取告警的详细统计数据
      parameters:
      - description: 项目 Key
        in: query
        name: project_key
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: 分组方式 (day/week/month)
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_reporting_dto.AlertStatsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警统计
      tags:
      - Report
  /api/v1/reports/dashboard:
    get:
      description: 获取仪表盘的综合统计数据
      parameters:
      - description: 项目 Key
        in: query
        name: project_key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DashboardStatsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取仪表盘统计
      tags:
      - Report
  /api/v1/reports/issues:
    get:
      description: 获取工单的详细统计数据
      parameters:
      - description: 项目 Key
        in: query
        name: project_key
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: 分组方式 (day/week/month)
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.IssueStatsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取工单统计
      tags:
      - Report
  /api/v1/reports/sla:
    get:
      description: 获取 SLA 统计报表
      parameters:
      - description: 项目 Key
        in: query
        name: project_key
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SLAReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取 SLA 报表
      tags:
      - Report
  /api/v1/reports/user-performance:
    get:
      description: 获取用户处理工单的绩效数据
      parameters:
      - description: 项目 Key
        in: query
        name: project_key
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserPerformanceResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取用户绩效
      tags:
      - Report
  /api/v1/reports/worklogs:
    get:
      description: 获取工时统计数据（按日/按人/按类型汇总）
      parameters:
      - description: 项目 Key
        in: query
        name: project_key
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.WorklogStatsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取工时统计
      tags:
      - Report
...
  /api/v1/requirement-pools:
    get:
      consumes:
      - application/json
      description: 获取需求池列表（支持分页和筛选）
      parameters:
      - description: 类型
        enum:
        - global
        - project
        in: query
        name: type
        type: string
      - description: 状态
        enum:
        - active
        - archived
        in: query
        name: status
        type: string
      - description: 项目ID
        in: query
        name: project_id
        type: integer
      - description: 负责人ID
        in: query
        name: owner_id
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取需求池列表
      tags:
      - RequirementPool
    post:
      consumes:
      - application/json
      description: 创建一个新的需求池
      parameters:
      - description: 创建需求池请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRequirementPoolRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RequirementPoolResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建需求池
      tags:
      - RequirementPool
  /api/v1/requirement-pools/{id}:
    delete:
      consumes:
      - application/json
      description: 删除需求池（软删除）
      parameters:
      - description: 需求池ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除需求池
      tags:
      - RequirementPool
    get:
      consumes:
      - application/json
      description: 根据ID获取需求池详情
      parameters:
      - description: 需求池ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RequirementPoolResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取需求池详情
      tags:
      - RequirementPool
    put:
      consumes:
      - application/json
      description: 更新需求池信息
      parameters:
      - description: 需求池ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新需求池请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRequirementPoolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新需求池
      tags:
      - RequirementPool
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer 令牌 (`Bearer <jwt>`) 或 Personal Access Token (`Bearer td_pat_xxx`)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"